fix: rc versioning with multiple other prereleases
There are probably still some edge cases with rc versioning, but this
change resolves those caused by missing commits (because i was filtering
from the last rc in some cases). Also i was relying on the semver spec
for sort but now there's a sort implementation that understands the rc
semantics.