feat: end-to-end templating
This change makes --template work with reading previous tags. It does
this by rendering the provided template with glob characters injected
into the template data.

The limitation here is that there must be a continuous, valid SemVer
version in any custom template. There is a fancier way to generate a
regex using the template, but this change just regexp-extracts the
SemVer string out of the tag.